Fixing StyleCop errors
