+++ src/Microsoft.Atlas.CommandLine/Abstractions/FileSystem.cs
@@ -1,19 +1,11 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
 using System.IO;
 using System.Linq;
 
 namespace Microsoft.Atlas.CommandLine.Abstractions
 {
-    public interface IFileSystem
-    {
-        string PathCombine(params string[] paths);
-
-        bool DirectoryExists(string path);
-
-        bool FileExists(string path);
-
-        TextReader OpenText(string path);
-    }
-
     public class FileSystem : IFileSystem
     {
         public string PathCombine(params string[] paths)
