+++ src/Microsoft.Atlas.CommandLine/Blueprints/IBlueprintManager.cs
@@ -8,6 +8,7 @@ namespace Microsoft.Atlas.CommandLine.Blueprints
     public interface IBlueprintManager
     {
         Task<IBlueprintPackage> GetBlueprintPackage(string blueprint);
+
         Task<IBlueprintPackage> GetBlueprintPackageDependency(IBlueprintPackage package, string blueprint);
     }
 }
