+++ src/Microsoft.Atlas.CommandLine/Execution/IWorkflowEngine.cs
@@ -1,4 +1,7 @@
-﻿using System.Collections.Generic;
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.Atlas.CommandLine.Models.Workflow;
