+++ src/Microsoft.Atlas.CommandLine/Execution/IWorkflowLoader.cs
@@ -1,3 +1,6 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
 using System;
 using System.IO;
 using Microsoft.Atlas.CommandLine.Blueprints;
